+++ R/compute_surf_refl.R
@@ -1,4 +1,4 @@
-#' Computes BRF based on outputs from PROSAIL & SZA
+#' Computes surface reflectance based on outputs from PROSAIL & SZA
 #'
 #' The direct and diffuse light are taken into account as proposed by:
 #' Francois et al. (2002) Conversion of 400-1100 nm vegetation albedo
@@ -12,10 +12,10 @@
 #' @param tts numeric. Solar zenith angle
 #' @param spec_atm_sensor list. direct & diffuse radiation for clear conditions
 #' @param skyl numeric. values for skyl
-#' @return BRF numeric. Bidirectional reflectance factor
+#' @return surf_refl numeric. surface reflectance
 #' @export
 
-compute_brf  <- function(rdot, rsot, tts, spec_atm_sensor, skyl = NULL){
+compute_surf_refl  <- function(rdot, rsot, tts, spec_atm_sensor, skyl = NULL){
 
   ############################## #
   ##	direct / diffuse light	##
@@ -28,14 +28,6 @@ compute_brf  <- function(rdot, rsot, tts, spec_atm_sensor, skyl = NULL){
     skyl <- 0.847-1.61*sin((90-tts)*rd) + 1.04*sin((90-tts)*rd)*sin((90-tts)*rd)
   par_diro <- (1-skyl)*es
   par_difo <- skyl*ed
-  brf <- (rdot*par_difo+rsot*par_diro)/(par_diro+par_difo)
-  return(data.frame('BRF' = brf))
-}
-
-
-#' @rdname prosail-deprecated
-#' @export
-Compute_BRF  <- function(rdot, rsot, tts, SpecATM_Sensor, skyl = NULL){
-  .Deprecated("compute_brf")
-  compute_brf(rdot, rsot, tts, SpecATM_Sensor, skyl)
+  surf_refl <- (rdot*par_difo+rsot*par_diro)/(par_diro+par_difo)
+  return(data.frame('surf_refl' = surf_refl))
 }