+++ R/get_atbd_LUT_input.R
@@ -68,12 +68,12 @@ get_atbd_lut_input <- function(nb_samples = 2000, geom_acq = NULL,
                                                   'soil_brightness' = 1.2)
     for (parm in names(codistribution_lai$vmin_0)){
       v_star <- get_codistributions(V = input_prosail[[parm]],
-                                   lai = input_prosail$lai,
-                                   max_lai = t_gauss_parms$max$lai,
-                                   vmin_0 = codistribution_lai$vmin_0[[parm]],
-                                   vmax_0 = codistribution_lai$vmax_0[[parm]],
-                                   vmin_lai_max = codistribution_lai$vmin_lai_max[[parm]],
-                                   vmax_lai_max = codistribution_lai$vmax_lai_max[[parm]])
+                                    lai = input_prosail$lai,
+                                    max_lai = t_gauss_parms$max$lai,
+                                    vmin_0 = codistribution_lai$vmin_0[[parm]],
+                                    vmax_0 = codistribution_lai$vmax_0[[parm]],
+                                    vmin_lai_max = codistribution_lai$vmin_lai_max[[parm]],
+                                    vmax_lai_max = codistribution_lai$vmax_lai_max[[parm]])
       input_prosail[[parm]] <- v_star
     }
   }
@@ -88,14 +88,14 @@ get_atbd_lut_input <- function(nb_samples = 2000, geom_acq = NULL,
   # set geometry of acquisition
   if (is.null(geom_acq))
     geom_acq <- data.frame('min' = c('tto' = 0, 'tts' = 20, 'psi' = 0),
-                          'max' = c('tto' = 10, 'tts' = 30, 'psi' = 360))
+                           'max' = c('tto' = 10, 'tts' = 30, 'psi' = 360))
   if (inherits(geom_acq, "list"))
     geom_acq <- data.frame('min' = c('tto' = geom_acq$min$tto,
-                                    'tts' = geom_acq$min$tts,
-                                    'psi' = geom_acq$min$psi),
-                          'max' = c('tto' = geom_acq$max$tto,
-                                    'tts' = geom_acq$max$tts,
-                                    'psi' = geom_acq$max$psi))
+                                     'tts' = geom_acq$min$tts,
+                                     'psi' = geom_acq$min$psi),
+                           'max' = c('tto' = geom_acq$max$tto,
+                                     'tts' = geom_acq$max$tts,
+                                     'psi' = geom_acq$max$psi))
   input_prosail$tts <- runif(n = nb_samples, min = geom_acq['tts', 'min'],
                              max = geom_acq['tts', 'max'])
   input_prosail$tto <- runif(n = nb_samples, min = geom_acq['tto', 'min'],