+++ R/get_spec_sensor.R
@@ -20,7 +20,7 @@ get_spec_sensor <- function(sensor_name = 'user_defined',
   if (is.null(srf)) stop()
   # adjust optical constants from 1 nm sampling to S2 spectral sampling
   spec_sensor <- prepare_sensor_simulation(spec_prospect = prosail::spec_prospect_full_range,
-                                           spec_soil = prosail::spec_soil_ossl,
+                                           spec_soil = prosail::spec_soil_atbd_v2,
                                            spec_atm = prosail::spec_atm,
                                            srf = srf)
   return(spec_sensor)