+++ tests/testthat/test-compute_surface_refl.R
@@ -1,10 +1,10 @@
-test_that("prosail produces physically possible BRF values", {
+test_that("prosail produces physically possible surf_refl values", {
   # run PROSAIL with 4SAIL2
   refl <- prosail()
-  brf_4sail <- compute_brf(rdot = refl$rdot,
-                           rsot = refl$rsot,
-                           tts = 30,
-                           spec_atm_sensor = spec_atm)
+  surf_refl_4sail <- compute_surf_refl(rdot = refl$rdot,
+                                       rsot = refl$rsot,
+                                       tts = 30,
+                                       spec_atm_sensor = spec_atm)
 
   # run PROSAIL with 4SAIL2
   input_prospect <- data.frame('chl' = c(40, 5), 'car' = c(8, 4),
@@ -15,11 +15,11 @@ test_that("prosail produces physically possible BRF values", {
                          type_lidf = 2, lidf_a = 30, lai = 5, hotspot = 0.1, tts = 30,
                          tto = 10, psi = 90, rsoil = spec_soil$max_refl,
                          fraction_brown = 0.5, diss = 0.5, cv = 1, zeta = 1)
-  brf_4sail2 <- compute_brf(rdot = refl_4sail2$rdot,
-                            rsot = refl_4sail2$rsot,
-                            tts = 30,
-                            spec_atm_sensor = spec_atm)
+  surf_refl_4sail2 <- compute_surf_refl(rdot = refl_4sail2$rdot,
+                                        rsot = refl_4sail2$rsot,
+                                        tts = 30,
+                                        spec_atm_sensor = spec_atm)
 
-  expect_true(all(brf_4sail$BRF >= 0))
-  expect_true(all(brf_4sail2$BRF >= 0))
+  expect_true(all(surf_refl_4sail$surf_refl >= 0))
+  expect_true(all(surf_refl_4sail2$surf_refl >= 0))
 })