+++ NEWS.md
@@ -20,6 +20,7 @@ prepare for JOSS submission
 - change 'get_radiometry' to 'get_spectral_response_function'
 - adjust range of ewt to avoid extremely high values when calling 'get_atbd_v3_lut_input'
 - add support for package prospect v2.0
+- use surface reflectance instead of brf
 
 ##fix 
 - fixed bug when introducing OSSL soil and disactivate OSSL for v3 beta